requirements.txt
kafka-python
pymongo
pandas
matplotlib
scikit-learn
pyflink
